feature: improves loading states for dashboard pages

+++ app/dashboard/explore/page.tsx
@@ -32,57 +32,64 @@ const ExploreEvents = () => {
   });
 
   const handleDelete = async (id: Key) => {
-    let filteredEvents = [...events];
-    filteredEvents = filteredEvents.filter(event => event.id !== id);
-    const response = await fetch(`http://localhost:3000/api/events/${id}`, {
-      method: 'DELETE',
-      body: JSON.stringify(id),
-    });
-    if (response.status === 204) {
-      setEvents(filteredEvents);
-      toast({
-        description: 'Event deleted.',
+    if (events) {
+      let filteredEvents = [...events];
+      filteredEvents = filteredEvents.filter(event => event.id !== id);
+      const response = await fetch(`http://localhost:3000/api/events/${id}`, {
+        method: 'DELETE',
+        body: JSON.stringify(id),
       });
-    } else {
-      toast({
-        title: 'Uh oh!',
-        description:
-          'There was an issue deleting your event. Try again in a few seconds.',
-      });
+      if (response.status === 204) {
+        setEvents(filteredEvents);
+        toast({
+          description: 'Event deleted.',
+        });
+      } else {
+        toast({
+          title: 'Uh oh!',
+          description:
+            'There was an issue deleting your event. Try again in a few seconds.',
+        });
+      }
     }
   };
+
+  // Events initialized as []
+  // Null means no events
+  // Empty events means initial loading state
+  // Filled events array means events were returned
 
   return (
     <div>
       <PageHeading>Explore</PageHeading>
       <PageSubHeading>Discover upcoming public events near you:</PageSubHeading>
       <CardGrid>
-        {errors && (
+        {events?.length === 0 && <p>Loading...</p>}
+        {errors ? (
           <div>
             <p>Uh oh!</p>
             <p>There was an issue retrieving your events.</p>
           </div>
-        )}
-        {events.length === 0 && !errors ? (
-          <p>We could not find any upcoming events near you.</p>
         ) : (
-          ''
+          <>
+            {events
+              ?.filter(item => new Date() < new Date(item.date))
+              .map(({ id, title, date, body, owned_by }) => {
+                return (
+                  <EventCard
+                    handleDelete={handleDelete}
+                    key={id as Key}
+                    id={id}
+                    title={title}
+                    date={date}
+                    body={body}
+                    isOwned={userID === owned_by ? true : false}
+                  />
+                );
+              })}
+          </>
         )}
-        {events
-          .filter(item => new Date() < new Date(item.date))
-          .map(({ id, title, date, body, owned_by }) => {
-            return (
-              <EventCard
-                handleDelete={handleDelete}
-                key={id as Key}
-                id={id}
-                title={title}
-                date={date}
-                body={body}
-                isOwned={userID === owned_by ? true : false}
-              />
-            );
-          })}
+        {!events && <p>We could not find any upcoming events near you.</p>}
       </CardGrid>
     </div>
   );
